# docs/openapi.yaml
openapi: 3.0.3
info:
  title: InfiAgentic Platform API
  description: |
    Complete API documentation for the InfiAgentic Marketing Automation Platform.
    
    ## Authentication
    All endpoints require JWT Bearer token authentication except `/auth/login` and `/auth/register`.
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - Premium: 1000 requests/minute
    
    ## Products
    - **Jyotishya** - Astrology SaaS
    - **HealthPedya** - Healthcare Portal
    - **SchooPay** - Education Payments
  version: 1.0.0
  contact:
    name: InfiDevelopers API Support
    email: api@infidevelopers.com
    url: https://infidevelopers.com/support
  license:
    name: Proprietary
    url: https://infidevelopers.com/license
servers:
  - url: https://api.infiagentic.com/v1
    description: Production
  - url: https://staging-api.infiagentic.com/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Local Development
...
security:
  - BearerAuth: []
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products
      responses:
        '200':
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

    post:
      tags: [Products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Product created
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'

    put:
      tags: [Products]
      summary: Update product
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Product updated
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        type:
          type: string
        description:
          type: string
        is_active:
          type: boolean
    ProductCreate:
      type: object
      required: [name, slug, type]
      properties:
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [jyotishya, healthpedya, schoopay, custom]
        description:
          type: string
        api_key:
          type: string
        webhook_url:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        settings:
          type: object